/*** Release1 ***/ - ack & nack verification - byte[] frame => int[] frame (signed to unsigned in java shit)

+++ src/Consumer.java
@@ -11,11 +11,11 @@ import java.util.concurrent.BlockingQueue;
 
 public class Consumer implements Runnable{
 
-    private BlockingQueue<byte[]> queue;
+    private BlockingQueue<int[]> queue;
     private Controller _ctrl;
     private Boolean stateFrame;
 
-    public Consumer(BlockingQueue<byte[]> q, Producer producer,  Controller controller){
+    public Consumer(BlockingQueue<int[]> q, Producer producer,  Controller controller){
         this._ctrl = controller;
         this.stateFrame = true;
         this.queue = q;
@@ -24,7 +24,7 @@ public class Consumer implements Runnable{
     public synchronized void run() {
         try{
             while(queue.isEmpty() && stateFrame){
-                byte[] frame = queue.take();
+                int[] frame = queue.take();
                 this.handleFrame(frame);
             }
         } catch(InterruptedException e) {
@@ -37,7 +37,7 @@ public class Consumer implements Runnable{
     public Boolean getStateFrame(){
         return this.stateFrame;
     }
-    private void handleFrame(byte[] frame){
+    private void handleFrame(int[] frame){
         System.out.print("Consumed : ");
         for (int i=0; i<21 ; i++){
                 System.out.print(frame[i] + " ");
